fixturize batch test
