+++ test/unit/ConfigurationTest.rb
