+++ test/unit/PaginatedArrayTest.rb
