+++ test/unit/ResponseMapperTest.rb
